Adding FBX models with animations
I'm including 3 FBX files in this folder:
"Reset Rat" is the one that worked for me when I dragged it into the assets and tried to play the animations, they all worked as intended.
I'm also including "Action Rat" if that works better with how animation is handled.
"RestingRat" is the default resting pose that all the animation is based on and relative to.